Proposed make system using docker-compose.
This system seems much easier to use:
make DESIGN=<design> will bring up all the containers and execute commands.
The dependencies are fully managed.

It's likely we can construct a command and execute in a shell instead if
our environment does not need docker containers (the environment is the union
of all container environments).

But first we should replace the existing make system in circleci with this one
as it can run multiple designs quickly as containers are live and this stress
tests the environments (e.g. for stale data, etc).