Simplified make system with flexible volume/dir output (Ready to Merge). (#131)
* Simplified make system with flexible volume/dir output

This version fixes several potential problems when changing volume output, as it
detects the change and rebuilds the containers.  Volumes can now be renamed (we
were stuck on compose_dataVolume previously).  Dependency checking is much better
and should bring up containers more consistently.  Documentation is better aligned
to this mode.

* Codacy fix of README.md [with change to Codacy rules]